fix(ground-truth): code-only /redteam /iq200 audit — fix 57 findings across 43 files
Core Python (46 findings):
- P0: Fix dead-code auth guard in a2a/server (identity.issuer→sub)
- P0: Fix PHI scrubber silently disabled on YAML parse error
- P0: Fix import-time crash in gcp/embedder (unguarded retry decorator)
- P1: Fix thread-safety in auth locks, OTel setup, token accumulator
- P1: Fix resource leaks (OTel context, DB cursor, fd double-close)
- P1: Fix SyntaxError in otel_setup.py setup_json_logging
- P1: Fix non-atomic writes in rejected.py (data corruption)
- P1: Fix 5th-judge tie-break bias (hardcoded safety axis)
- P1: Fix missing timeouts in hermes_bridge subprocess
- P2: Fix type annotations, input validation, error handling

Infrastructure (9 findings):
- P0: Fix silent :latest fallback in GCP compose override
- P0: Fix unverified cosign download in deploy workflow
- P0: Fix missing set -e in smoke.sh
- P1: Downscope IAM roles (objectAdmin→objectCreator)
- P1: Add prevent_destroy to forensic log bucket
- P1: Fix unquoted paths in panic.sh/teardown.sh

Cross-cutting (2 findings):
- P0: Fix unguarded google-cloud-aiplatform import
- P1: Fix dead-code auth predicate (algebraically unreachable)

+++ lib/durability/trichotomy.py
@@ -133,7 +133,7 @@ def after_tool_call(
     task_id: Optional[str] = None,
     session_id: Optional[str] = None,
     tool_call_id: Optional[str] = None,
-    duration_ms: Optional[int] = None,
+    duration_ms: Optional[float] = None,
     **_: Any,
 ) -> None:
     """Hermes ``post_tool_call`` hook. Classifies errors, emits span, dispatches handler.
